==> Cleaning Scripts/00_packages.R <==
# Load packages -----------------------------------------------------------

library(changepoint)
library(cowplot)
library(geosphere)
library(ggplot2)
library(ggpubr)
library(ggridges)
library(ICSNP)
library(igraph)
library(kableExtra)
library(lubridate)
library(MASS)
library(moments)
library(mrfDepth)
library(patchwork)
library(POT)
library(tidyverse)
library(zoo)
library(extraDistr)
library(fdapace)
library(httr)
library(jsonlite)


# Load functions ----------------------------------------------------------

source("Functions/matrix_functions.R")
source("Functions/notin.R")
source("Functions/aggregate_start_end.R")
source("Functions/correlation_matrix_function.R")
source("Functions/weather_api_function.R")
source("Functions/date_format.R")
source("Functions/clustering.R")
source("Functions/sse_function.R.R")
source("Functions/residuals_function_model_comp.R")
source("Functions/wrapper_function.R")
source("Functions/wrapper_log_function.R")
source("Functions/residuals_function.R")
source("Functions/mst_clustering.R")
source("Functions/depth.R")
source("Functions/depth_threshold.R")
source("Functions/merge_differences.R")
source("Functions/beta4_probs.R")
source("Functions/name_rem_x.R")
source("Functions/season_split.R")
source("Functions/day_split.R")
source("Functions/estBetaParams.R")
source("Functions/day_season_split.R")
